add manpage.rb to generating homebrew manual
