+++ Library/Homebrew/manpages.rbi
@@ -1,0 +1,7 @@
+# typed: strict
+
+module Homebrew
+  module Manpages
+    include Kernel
+  end
+end
